fix(navbar): 로그인 완료 후 내 정보 제대로 받아오고 navbar에 표시 #120
