+++ src/pages/FrontPage.jsx
@@ -1,26 +1,16 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { Hstack } from "../package/layout";
 import CustomButton from "../package/customButton/CustomButton";
 import RecommendContent from "../features/front/recommend/RecommendContent";
 import TotalContent from "../features/front/total/TotalContent";
 import SearchContent from "../features/front/search/SearchContent";
 import { useFront } from "../features/front/useFront";
-import useLinkUpStore from "../shared/store/store";
-import { useSearchParams } from "react-router";
 
 /** 임시로 각 콘텐트로 이동시키게 만들었습니다. */
 const FrontPage = () => {
     const [whatToShow, setWhatToShow] = useState(null);
-    const [searchParams, _setSearchParams] = useSearchParams();
-    const setAccessToken = useLinkUpStore((state) => state.setAccessToken);
 
     useFront();
-
-    useEffect(() => {
-        const access_token = searchParams.get("access_token");
-        console.log({ access_token });
-        setAccessToken(access_token);
-    }, []);
 
     if (!whatToShow) {
         return (
